incorporate review comments.

+++ projects/rocshmem/tests/functional_tests/team_broadcast_tester.cpp
@@ -221,7 +221,7 @@ void TeamBroadcastTester<T1>::verifyResults(size_t size) {
    * the broadcast.
    */
   for (unsigned int wg_id = 0; wg_id < args.num_wgs; wg_id++) {
-    for (unsigned int i = 0; i < static_cast<unsigned int>(num_elems); i++) {
+    for (int i = 0; i < num_elems; i++) {
       idx = wg_id * num_elems + static_cast<int>(i);
       if constexpr (std::is_same<T1, char>::value ||
                     std::is_same<T1, signed char>::value ||
